[build] devide repository
