+++ src/feature/api/channels/domain/channel.ts
@@ -3,6 +3,7 @@ export class Channel {
   private readonly _name: string;
   private readonly _status: string;
   private readonly _password: string;
+  private readonly _isDeleted: boolean;
 
   // TODO: implement validation
   constructor(props: {
@@ -10,11 +11,13 @@ export class Channel {
     name: string;
     status: string;
     password: string;
+    isDeleted: boolean;
   }) {
     this._id = props.id;
     this._name = props.name;
     this._status = props.status;
     this._password = props.password;
+    this._isDeleted = props.isDeleted;
   }
 
   get id(): string {
@@ -31,5 +34,9 @@ export class Channel {
 
   get password(): string {
     return this._password;
+  }
+
+  get isDeleted(): boolean {
+    return this._isDeleted;
   }
 }
